@@ -1,6 +1,6 @@
 {
   "name": "tauri",
-  "version": "0.3.1",
+  "version": "0.4.5",
   "description": "Multi-binding collection of libraries and templates for building Tauri apps",
   "bin": {
     "tauri": "./bin/tauri.js"
@@ -11,9 +11,10 @@
   },
   "scripts": {
     "build": "webpack --progress",
+    "build-release": "yarn build --display none --progress false",
     "test": "jest --runInBand --no-cache",
-    "prepare": "yarn build",
     "pretest": "yarn build",
+    "prepublishOnly": "yarn build-release",
     "test:mac-local": "jest --runInBand",
     "lint": "eslint --ext ts ./src/**/*.ts",
     "lint-fix": "eslint --fix --ext ts ./src/**/*.ts",
@@ -43,12 +44,12 @@
     "yarn": ">= 1.19.1"
   },
   "dependencies": {
-    "@tauri-apps/tauri-inliner": "1.13.2",
+    "@tauri-apps/tauri-inliner": "1.14.0",
     "@tauri-apps/toml": "2.2.4",
     "chalk": "3.0.0",
     "chokidar": "3.3.1",
     "cross-spawn": "7.0.1",
-    "fast-glob": "3.1.1",
+    "fast-glob": "3.2.2",
     "fs-extra": "8.1.0",
     "imagemin": "7.0.1",
     "imagemin-optipng": "7.1.0",
@@ -56,13 +57,13 @@
     "imagemin-zopfli": "6.0.0",
     "is-png": "2.0.0",
     "isbinaryfile": "4.0.4",
-    "jsdom": "16.0.1",
+    "jsdom": "16.2.0",
     "lodash": "4.17.15",
     "minimist": "1.2.0",
     "ms": "2.1.2",
     "png2icons": "2.0.1",
     "read-chunk": "3.2.0",
-    "sharp": "0.23.4",
+    "sharp": "0.24.1",
     "webpack-merge": "4.2.2",
     "webpack-shell-plugin": "0.5.0"
   },
@@ -71,38 +72,38 @@
     "@babel/preset-env": "7.7.7",
     "@babel/preset-typescript": "7.7.7",
     "@types/cross-spawn": "6.0.1",
-    "@types/fs-extra": "8.0.1",
+    "@types/fs-extra": "8.1.0",
     "@types/imagemin": "7.0.0",
     "@types/imagemin-optipng": "5.2.0",
-    "@types/jsdom": "12.2.4",
+    "@types/jsdom": "16.1.0",
     "@types/lodash": "4.14.149",
     "@types/ms": "0.7.31",
     "@types/sharp": "0.24.0",
     "@types/webpack-merge": "4.1.5",
-    "@typescript-eslint/eslint-plugin": "2.17.0",
-    "@typescript-eslint/parser": "2.17.0",
+    "@typescript-eslint/eslint-plugin": "2.22.0",
+    "@typescript-eslint/parser": "2.22.0",
     "babel-jest": "24.9.0",
     "dotenv": "8.2.0",
     "eslint": "6.8.0",
-    "eslint-config-standard-with-typescript": "12.0.1",
-    "eslint-plugin-import": "2.20.0",
+    "eslint-config-standard-with-typescript": "14.0.0",
+    "eslint-plugin-import": "2.20.1",
     "eslint-plugin-lodash-template": "0.15.0",
     "eslint-plugin-node": "11.0.0",
     "eslint-plugin-promise": "4.2.1",
     "eslint-plugin-security": "1.4.0",
     "eslint-plugin-standard": "4.0.1",
-    "husky": "4.2.0",
+    "husky": "4.2.3",
     "is-running": "^2.1.0",
     "jest": "25.1.0",
-    "jest-mock-process": "1.3.0",
-    "lint-staged": "10.0.2",
-    "lockfile-lint": "3.0.8",
-    "promise": "8.0.3",
+    "jest-mock-process": "1.3.2",
+    "lint-staged": "10.0.8",
+    "lockfile-lint": "4.0.0",
+    "promise": "8.1.0",
     "raw-loader": "4.0.0",
     "ts-loader": "6.2.1",
-    "typescript": "3.7.5",
-    "webpack": "4.41.5",
-    "webpack-cli": "3.3.10",
+    "typescript": "3.8.3",
+    "webpack": "4.42.0",
+    "webpack-cli": "3.3.11",
     "webpack-node-externals": "1.7.2"
   },
   "husky": {
@@ -111,7 +112,6 @@
     }
   },
   "lint-staged": [
-    "eslint --fix",
-    "git add"
+    "eslint --fix"
   ]
 }
