@@ -3,7 +3,8 @@ const parseArgs = require('minimist')
 const argv = parseArgs(process.argv.slice(2), {
   alias: {
     h: 'help',
-    d: 'debug'
+    d: 'debug',
+    t: 'target'
   },
   boolean: ['h', 'd']
 })
@@ -16,10 +17,17 @@ if (argv.help) {
     $ tauri build
   Options
     --help, -h     Displays this message
+    --debug, -d    Builds with the debug flag
+    --target, -t   Comma-separated list of target triples to build against
   `)
   process.exit(0)
 }
 
 const build = require('../dist/api/build')
 
-build({ ctx: { debug: argv.debug } })
+build({
+  ctx: {
+    debug: argv.debug,
+    target: argv.target
+  }
+})
