@@ -7,7 +7,7 @@ import * as appPaths from './app-paths'
 
 const error = logger('ERROR:', 'red')
 
-module.exports = (cfg: Partial<TauriConfig>): TauriConfig => {
+const getTauriConfig = (cfg: Partial<TauriConfig>): TauriConfig => {
   const pkgPath = appPaths.resolve.app('package.json')
   const tauriConfPath = appPaths.resolve.tauri('tauri.conf.json')
   if (!existsSync(pkgPath)) {
@@ -46,6 +46,9 @@ module.exports = (cfg: Partial<TauriConfig>): TauriConfig => {
         },
         edge: {
           active: true
+        },
+        inliner: {
+          active: true
         }
       }
     } as any,
@@ -67,3 +70,5 @@ module.exports = (cfg: Partial<TauriConfig>): TauriConfig => {
 
   return config
 }
+
+export default getTauriConfig
