@@ -3,7 +3,7 @@ import toml from '@tauri-apps/toml'
 import chokidar, { FSWatcher } from 'chokidar'
 import { existsSync, readFileSync, writeFileSync } from 'fs-extra'
 import { JSDOM } from 'jsdom'
-import { debounce } from 'lodash'
+import { debounce, template } from 'lodash'
 import path from 'path'
 import * as entry from './entry'
 import { appDir, tauriDir } from './helpers/app-paths'
@@ -79,7 +79,7 @@ class Runner {
           path.join(tauriDir, 'src'),
           path.join(tauriDir, 'Cargo.toml'),
           path.join(tauriDir, 'build.rs'),
-          path.join(appDir, 'tauri.conf.js')
+          path.join(appDir, 'tauri.conf.json')
         ],
         {
           ignoreInitial: true
@@ -90,7 +90,7 @@ class Runner {
         debounce((path: string) => {
           this.__stopCargo()
             .then(() => {
-              if (path.includes('tauri.conf.js')) {
+              if (path.includes('tauri.conf.json')) {
                 this.run(getTauriConfig({ ctx: cfg.ctx })).catch(e => {
                   throw e
                 })
@@ -127,7 +127,7 @@ class Runner {
     const buildFn = async (target?: string): Promise<void> =>
       this.__runCargoCommand({
         cargoArgs: [
-          cfg.tauri.bundle.active ? 'tauri-cli' : 'build',
+          cfg.tauri.bundle.active ? 'tauri-bundler' : 'build',
           '--features',
           ...features
         ]
@@ -135,8 +135,8 @@ class Runner {
           .concat(target ? ['--target', target] : [])
       })
 
-    if (cfg.ctx.debug || !cfg.ctx.targetName) {
-      // on debug mode or if no target specified,
+    if (!cfg.ctx.target) {
+      // if no target specified,
       // build only for the current platform
       await buildFn()
     } else {
@@ -167,6 +167,25 @@ class Runner {
           interceptor(dom)
         }
 
+        if (!((cfg.ctx.dev && cfg.build.devPath.startsWith('http')) || cfg.tauri.embeddedServer.active)) {
+          const mutationObserverTemplate = require('!!raw-loader!!../templates/mutation-observer').default
+          const compiledMutationObserver = template(mutationObserverTemplate)
+
+          const bodyMutationObserverScript = document.createElement('script')
+          bodyMutationObserverScript.text = compiledMutationObserver({
+            target: 'body',
+            inlinedAssets: JSON.stringify(inlinedAssets)
+          })
+          document.body.insertBefore(bodyMutationObserverScript, document.body.firstChild)
+
+          const headMutationObserverScript = document.createElement('script')
+          headMutationObserverScript.text = compiledMutationObserver({
+            target: 'head',
+            inlinedAssets: JSON.stringify(inlinedAssets)
+          })
+          document.head.insertBefore(headMutationObserverScript, document.head.firstChild)
+        }
+
         const tauriScript = document.createElement('script')
         // @ts-ignore
         tauriScript.text = readFileSync(path.join(tauriDir, 'tauri.js'))
@@ -185,21 +204,23 @@ class Runner {
         )
       }
 
-      if (cfg.tauri.embeddedServer.active) {
-        rewriteHtml(readFileSync(indexPath).toString())
+      const domInterceptor = cfg.tauri.embeddedServer.active ? undefined : (dom: JSDOM) => {
+        const document = dom.window.document
+        document.querySelectorAll('link').forEach((link: HTMLLinkElement) => {
+          link.removeAttribute('rel')
+          link.removeAttribute('as')
+        })
+      }
+
+      if (cfg.tauri.embeddedServer.active || !cfg.tauri.inliner.active) {
+        rewriteHtml(readFileSync(indexPath).toString(), domInterceptor)
         resolve(inlinedAssets)
       } else {
         new Inliner(indexPath, (err: Error, html: string) => {
           if (err) {
             reject(err)
           } else {
-            rewriteHtml(html, dom => {
-              const document = dom.window.document
-              document.querySelectorAll('link').forEach(link => {
-                link.removeAttribute('rel')
-                link.removeAttribute('as')
-              })
-            })
+            rewriteHtml(html, domInterceptor)
             resolve(inlinedAssets)
           }
         }).on('progress', (event: string) => {
@@ -256,6 +277,8 @@ class Runner {
             warn()
             process.exit(0)
           }
+
+          resolve()
         }
       )
       
