@@ -18,6 +18,40 @@
  * and also whitelist them based upon the developer's settings.
  */
 
+// polyfills
+if (!String.prototype.startsWith) {
+  String.prototype.startsWith = function (searchString, position) {
+    position = position || 0
+    return this.substr(position, searchString.length) === searchString
+  }
+}
+
+// makes the window.external.invoke API available after window.location.href changes
+
+switch (navigator.platform) {
+  case "Macintosh":
+  case "MacPPC":
+  case "MacIntel":
+  case "Mac68K":
+    window.external = this
+    invoke = function (x) {
+      webkit.messageHandlers.invoke.postMessage(x);
+    }
+    break;
+  case "Windows":
+  case "WinCE":
+  case "Win32":
+  case "Win64":
+    break;
+  default: 
+    window.external = this
+    invoke = function (x) {
+      window.webkit.messageHandlers.external.postMessage(x);
+    }
+    break;
+}
+
+
 function s4() {
   return Math.floor((1 + Math.random()) * 0x10000)
     .toString(16)
@@ -47,10 +81,10 @@ function _typeof(obj) { if (typeof Symbol === "function" && typeof Symbol.iterat
  * @private
  */
 var __whitelistWarning = function (func) {
-  console.warn('%c[Tauri] Danger \ntauri.' + func + ' not whitelisted 💣\n%c\nAdd to tauri.conf.json: \n\ntauri: \n  whitelist: { \n    ' + func + ': true \n\nReference: https://github.com/tauri-apps/tauri/wiki' + func , 'background: red; color: white; font-weight: 800; padding: 2px; font-size:1.5em', ' ')
-  return __reject()
-}
-<% } %>
+    console.warn('%c[Tauri] Danger \ntauri.' + func + ' not whitelisted 💣\n%c\nAdd to tauri.conf.json: \n\ntauri: \n  whitelist: { \n    ' + func + ': true \n\nReference: https://github.com/tauri-apps/tauri/wiki' + func, 'background: red; color: white; font-weight: 800; padding: 2px; font-size:1.5em', ' ')
+    return __reject()
+  }
+    <% } %>
 
 <% if (ctx.dev) { %>
   /**
@@ -92,18 +126,18 @@ window.tauri = {
   listen: function listen(event, handler) {
     <% if (tauri.whitelist.event === true || tauri.whitelist.all === true) { %>
     var once = arguments.length > 2 && arguments[2] !== undefined ? arguments[2] : false;
-    this.invoke({
-      cmd: 'listen',
-      event: event,
-      handler: window.tauri.transformCallback(handler, once),
-      once: once
-    });
+      this.invoke({
+        cmd: 'listen',
+        event: event,
+        handler: window.tauri.transformCallback(handler, once),
+        once: once
+      });
     <% } else { %>
       <% if (ctx.dev) { %>
           return __whitelistWarning('event')
           <% } %>
         return __reject()
-    <% } %>
+        <% } %>
   },
 
   <% if (ctx.dev) { %>
@@ -116,17 +150,17 @@ window.tauri = {
   <% } %>
   emit: function emit(evt, payload) {
     <% if (tauri.whitelist.event === true || tauri.whitelist.all === true) { %>
-    this.invoke({
-      cmd: 'emit',
-      event: evt,
-      payload: payload || ''
-    });
+      this.invoke({
+        cmd: 'emit',
+        event: evt,
+        payload: payload || ''
+      });
     <% } else { %>
       <% if (ctx.dev) { %>
           return __whitelistWarning('event')
           <% } %>
         return __reject()
-    <% } %>
+        <% } %>
   },
 
   <% if (ctx.dev) { %>
@@ -140,7 +174,7 @@ window.tauri = {
   <% } %>
   transformCallback: function transformCallback(callback) {
     var once = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : false;
-    var identifier = Object.freeze(uid());
+    var identifier = uid();
 
     window[identifier] = function (result) {
       if (once) {
@@ -183,17 +217,16 @@ window.tauri = {
   <% } %>
   readTextFile: function readTextFile(path) {
     <% if (tauri.whitelist.readTextFile === true || tauri.whitelist.all === true) { %>
-    Object.freeze(path);
-    return this.promisified({
-      cmd: 'readTextFile',
-      path: path
-    });
+      return this.promisified({
+        cmd: 'readTextFile',
+        path: path
+      });
     <% } else { %>
       <% if (ctx.dev) { %>
           return __whitelistWarning('readTextFile')
           <% } %>
         return __reject()
-    <% } %>
+        <% } %>
   },
 
   <% if (ctx.dev) { %>
@@ -207,17 +240,16 @@ window.tauri = {
   <% } %>
   readBinaryFile: function readBinaryFile(path) {
     <% if (tauri.whitelist.readBinaryFile === true || tauri.whitelist.all === true) { %>
-    Object.freeze(path);
-    return this.promisified({
-      cmd: 'readBinaryFile',
-      path: path
-    });
+      return this.promisified({
+        cmd: 'readBinaryFile',
+        path: path
+      });
     <% } else { %>
       <% if (ctx.dev) { %>
           return __whitelistWarning('readBinaryFile')
           <% } %>
         return __reject()
-    <% } %>
+        <% } %>
   },
 
   <% if (ctx.dev) { %>
@@ -232,18 +264,20 @@ window.tauri = {
   <% } %>
   writeFile: function writeFile(cfg) {
     <% if (tauri.whitelist.writeFile === true || tauri.whitelist.all === true) { %>
-    Object.freeze(cfg);
-    this.invoke({
-      cmd: 'writeFile',
-      file: cfg.file,
-      contents: cfg.contents
-    });
+      if (_typeof(cfg) === 'object') {
+        Object.freeze(cfg);
+      }
+      this.invoke({
+        cmd: 'writeFile',
+        file: cfg.file,
+        contents: cfg.contents
+      });
     <% } else { %>
       <% if (ctx.dev) { %>
           return __whitelistWarning('writeFile')
           <% } %>
         return __reject()
-    <% } %>
+        <% } %>
   },
 
   <% if (ctx.dev) { %>
@@ -257,18 +291,16 @@ window.tauri = {
   <% } %>
   listFiles: function listFiles(path) {
     <% if (tauri.whitelist.listFiles === true || tauri.whitelist.all === true) { %>
-
-    Object.freeze(path);
-    return this.promisified({
-      cmd: 'listFiles',
-      path: path
-    });
+      return this.promisified({
+        cmd: 'listFiles',
+        path: path
+      });
     <% } else { %>
       <% if (ctx.dev) { %>
           return __whitelistWarning('listDirs')
           <% } %>
         return __reject()
-    <% } %>
+        <% } %>
   },
 
   <% if (ctx.dev) { %>
@@ -282,17 +314,16 @@ window.tauri = {
   <% } %>
   listDirs: function listDirs(path) {
     <% if (tauri.whitelist.listDirs === true || tauri.whitelist.all === true) { %>
-    Object.freeze(path);
-    return this.promisified({
-      cmd: 'listDirs',
-      path: path
-    });
+      return this.promisified({
+        cmd: 'listDirs',
+        path: path
+      });
     <% } else { %>
       <% if (ctx.dev) { %>
           return __whitelistWarning('listDirs')
           <% } %>
         return __reject()
-    <% } %>
+        <% } %>
   },
 
   <% if (ctx.dev) { %>
@@ -304,17 +335,16 @@ window.tauri = {
   <% } %>
   setTitle: function setTitle(title) {
     <% if (tauri.whitelist.setTitle === true || tauri.whitelist.all === true) { %>
-    Object.freeze(title);
-    this.invoke({
-      cmd: 'setTitle',
-      title: title
-    });
+      this.invoke({
+        cmd: 'setTitle',
+        title: title
+      });
     <% } else { %>
       <% if (ctx.dev) { %>
     return __whitelistWarning('setTitle')
-      <% } %>
+          <% } %>
     return __reject()
-    <% } %>
+        <% } %>
   },
 
   <% if (ctx.dev) { %>
@@ -326,17 +356,16 @@ window.tauri = {
   <% } %>
   open: function open(uri) {
     <% if (tauri.whitelist.open === true || tauri.whitelist.all === true) { %>
-    Object.freeze(uri);
-    this.invoke({
-      cmd: 'open',
-      uri: uri
-    });
+      this.invoke({
+        cmd: 'open',
+        uri: uri
+      });
     <% } else { %>
     <% if (ctx.dev) { %>
       return __whitelistWarning('open')
-      <% } %>
+          <% } %>
     return __reject()
-      <% } %>
+        <% } %>
   },
 
   <% if (ctx.dev) { %>
@@ -352,31 +381,27 @@ window.tauri = {
   execute: function execute(command, args) {
     <% if (tauri.whitelist.execute === true || tauri.whitelist.all === true) { %>
 
-    Object.freeze(command);
+      if (_typeof(args) === 'object') {
+        Object.freeze(args);
+      }
 
-    if (typeof args === 'string' || _typeof(args) === 'object') {
-      Object.freeze(args);
-    }
-
-    return this.promisified({
-      cmd: 'execute',
-      command: command,
-      args: typeof args === 'string' ? [args] : args
-    });
+      return this.promisified({
+        cmd: 'execute',
+        command: command,
+        args: typeof args === 'string' ? [args] : args
+      });
     <% } else { %>
       <% if (ctx.dev) { %>
         return __whitelistWarning('execute')
-        <% } %>
+          <% } %>
         return __reject()
-    <% } %>
+        <% } %>
   },
 
-  bridge: function bridge(command, payload) {
+bridge: function bridge(command, payload) {
     <% if (tauri.whitelist.bridge === true || tauri.whitelist.all === true) { %>
 
-    Object.freeze(command);
-
-    if (typeof payload === 'string' || _typeof(payload) === 'object') {
+    if (_typeof(payload) === 'object') {
       Object.freeze(payload);
     }
 
@@ -388,18 +413,18 @@ window.tauri = {
     <% } else { %>
       <% if (ctx.dev) { %>
           return __whitelistWarning('bridge')
-      <% } %>
+        <% } %>
             return __reject()
-    <% } %>
+      <% } %>
   },
 
-  loadAsset: function loadAsset(assetName, assetType) {
-    return this.promisified({
-      cmd: 'loadAsset',
-      asset: assetName,
-      asset_type: assetType || 'unknown'
-    })
-  }
+loadAsset: function loadAsset(assetName, assetType) {
+  return this.promisified({
+    cmd: 'loadAsset',
+    asset: assetName,
+    asset_type: assetType || 'unknown'
+  })
+}
 };
 
 // init tauri API
@@ -420,7 +445,7 @@ document.addEventListener('error', function (e) {
   while (target != null) {
     if (target.matches ? target.matches('img') : target.msMatchesSelector('img')) {
       window.tauri.loadAsset(target.src, 'image')
-        .then(img => {
+        .then(function (img) {
           target.src = img
         })
       break
@@ -429,16 +454,27 @@ document.addEventListener('error', function (e) {
   }
 }, true)
 
-window.addEventListener('DOMContentLoaded', function () {
-  // open <a href="..."> links with the Tauri API
+// open <a href="..."> links with the Tauri API
+function __openLinks() {
   document.querySelector('body').addEventListener('click', function (e) {
     var target = e.target
     while (target != null) {
       if (target.matches ? target.matches('a') : target.msMatchesSelector('a')) {
-        window.tauri.open(target.href)
+        if (target.href && target.href.startsWith('http') && target.target === '_blank') {
+          window.tauri.open(target.href)
+          e.preventDefault()
+        }
         break
       }
       target = target.parentElement
     }
   }, true)
-}, true)
+}
+
+if (document.readyState === 'complete' || document.readyState === 'interactive') {
+  __openLinks()
+} else {
+  window.addEventListener('DOMContentLoaded', function () {
+    __openLinks()
+  }, true)
+}
