@@ -1,11 +1,35 @@
 use std::process::{Child, Command, Stdio};
 
+#[cfg(windows)]
+use std::os::windows::process::CommandExt;
+
+#[cfg(windows)]
+const CREATE_NO_WINDOW: u32 = 0x08000000;
+
 use tauri_utils::platform;
 
+#[cfg(not(windows))]
 pub fn get_output(cmd: String, args: Vec<String>, stdout: Stdio) -> crate::Result<String> {
   Command::new(cmd)
     .args(args)
     .stdout(stdout)
+    .output()
+    .map_err(|err| crate::Error::with_chain(err, "Command: get output failed"))
+    .and_then(|output| {
+      if output.status.success() {
+        Ok(String::from_utf8_lossy(&output.stdout).to_string())
+      } else {
+        Err(crate::ErrorKind::Command(String::from_utf8_lossy(&output.stderr).to_string()).into())
+      }
+    })
+}
+
+#[cfg(windows)]
+pub fn get_output(cmd: String, args: Vec<String>, stdout: Stdio) -> crate::Result<String> {
+  Command::new(cmd)
+    .args(args)
+    .stdout(stdout)
+    .creation_flags(CREATE_NO_WINDOW)
     .output()
     .map_err(|err| crate::Error::with_chain(err, "Command: get output failed"))
     .and_then(|output| {
@@ -27,23 +51,44 @@ pub fn format_command(path: String, command: String) -> String {
 
 pub fn relative_command(command: String) -> crate::Result<String> {
   match std::env::current_exe()?.parent() {
-    Some(exe_dir) => return Ok(format_command(exe_dir.display().to_string(), command)),
-    None => {
-      return Err(crate::ErrorKind::Command("Could not evaluate executable dir".to_string()).into())
-    }
+    Some(exe_dir) => Ok(format_command(exe_dir.display().to_string(), command)),
+    None => Err(crate::ErrorKind::Command("Could not evaluate executable dir".to_string()).into()),
   }
 }
 
+#[cfg(not(windows))]
 pub fn command_path(command: String) -> crate::Result<String> {
   match std::env::current_exe()?.parent() {
-    #[cfg(not(windows))]
     Some(exe_dir) => Ok(format!("{}/{}", exe_dir.display().to_string(), command)),
-    #[cfg(windows)]
+    None => Err(crate::ErrorKind::Command("Could not evaluate executable dir".to_string()).into()),
+  }
+}
+
+#[cfg(windows)]
+pub fn command_path(command: String) -> crate::Result<String> {
+  match std::env::current_exe()?.parent() {
     Some(exe_dir) => Ok(format!("{}/{}.exe", exe_dir.display().to_string(), command)),
     None => Err(crate::ErrorKind::Command("Could not evaluate executable dir".to_string()).into()),
   }
 }
 
+#[cfg(windows)]
+pub fn spawn_relative_command(
+  command: String,
+  args: Vec<String>,
+  stdout: Stdio,
+) -> crate::Result<Child> {
+  let cmd = relative_command(command)?;
+  Ok(
+    Command::new(cmd)
+      .args(args)
+      .creation_flags(CREATE_NO_WINDOW)
+      .stdout(stdout)
+      .spawn()?,
+  )
+}
+
+#[cfg(not(windows))]
 pub fn spawn_relative_command(
   command: String,
   args: Vec<String>,
@@ -54,7 +99,7 @@ pub fn spawn_relative_command(
 }
 
 pub fn binary_command(binary_name: String) -> crate::Result<String> {
-  return Ok(format!("{}-{}", binary_name, platform::target_triple()?));
+  Ok(format!("{}-{}", binary_name, platform::target_triple()?))
 }
 
 // tests for the commands functions.
