@@ -3,17 +3,13 @@
   windows_subsystem = "windows"
 )]
 
-#[cfg(test)]
-extern crate quickcheck;
-#[cfg(test)]
-#[macro_use(quickcheck)]
-extern crate quickcheck_macros;
-
 pub mod command;
 pub mod dir;
 pub mod file;
 pub mod rpc;
 pub mod version;
+
+pub use tauri_utils::*;
 
 use error_chain::error_chain;
 
