@@ -1,6 +1,6 @@
 pub fn format_callback(function_name: String, arg: String) -> String {
   let formatted_string = &format!("window[\"{}\"]({})", function_name, arg);
-  return formatted_string.to_string();
+  formatted_string.to_string()
 }
 
 pub fn format_callback_result(
@@ -9,14 +9,15 @@ pub fn format_callback_result(
   error_callback: String,
 ) -> String {
   match result {
-    Ok(res) => return format_callback(callback, res),
-    Err(err) => return format_callback(error_callback, format!("\"{}\"", err)),
+    Ok(res) => format_callback(callback, res),
+    Err(err) => format_callback(error_callback, format!("\"{}\"", err)),
   }
 }
 
 #[cfg(test)]
 mod test {
   use crate::rpc::*;
+  use quickcheck_macros::quickcheck;
 
   // check abritrary strings in the format callback function
   #[quickcheck]
