@@ -3,6 +3,7 @@ use std::fs;
 use std::path::PathBuf;
 
 use crate::http;
+
 use tauri_api::file::{Extract, Move};
 
 pub mod github;
@@ -243,7 +244,7 @@ impl Update {
 
     self.println("Downloading...");
     http::download(
-      &self.release.download_url,
+      self.release.download_url.clone(),
       &mut tmp_archive,
       self.show_download_progress,
     )?;
