@@ -10,7 +10,7 @@ pub fn get_latest_release(repo_owner: &str, repo_name: &str) -> crate::Result<Re
     "https://api.github.com/repos/{}/{}/releases/latest",
     repo_owner, repo_name
   );
-  let resp = http::get(&api_url)?;
+  let resp = http::get(api_url.clone())?;
   if !resp.status().is_success() {
     bail!(
       crate::ErrorKind::Network,
@@ -29,7 +29,7 @@ pub fn get_release_version(repo_owner: &str, repo_name: &str, ver: &str) -> crat
     "https://api.github.com/repos/{}/{}/releases/tags/{}",
     repo_owner, repo_name, ver
   );
-  let resp = http::get(&api_url)?;
+  let resp = http::get(api_url.clone())?;
   if !resp.status().is_success() {
     bail!(
       crate::ErrorKind::Network,
