@@ -1,3 +1,5 @@
+use std::path::{PathBuf, MAIN_SEPARATOR};
+
 /// Try to determine the current target triple.
 ///
 /// Returns a target triple (e.g. `x86_64-unknown-linux-gnu` or `i686-pc-windows-msvc`) or an
@@ -32,9 +34,8 @@ pub fn target_triple() -> Result<String, crate::Error> {
     return Err(crate::Error::from("Unable to determine target-os"));
   };
 
-  let s;
   let os = if cfg!(target_os = "macos") || cfg!(target_os = "freebsd") {
-    os
+    String::from(os)
   } else {
     let env = if cfg!(target_env = "gnu") {
       "gnu"
@@ -45,9 +46,41 @@ pub fn target_triple() -> Result<String, crate::Error> {
     } else {
       return Err(crate::Error::from("Unable to determine target-environment"));
     };
-    s = format!("{}-{}", os, env);
-    &s
+
+    format!("{}-{}", os, env)
   };
 
   Ok(format!("{}-{}", arch, os))
+}
+
+pub fn resource_dir() -> crate::Result<PathBuf> {
+  let exe = std::env::current_exe()?;
+  let exe_dir = exe.parent().expect("failed to get exe directory");
+  let app_name = exe
+    .file_name()
+    .expect("failed to get exe filename")
+    .to_string_lossy();
+  let curr_dir = exe_dir.display().to_string();
+
+  if curr_dir.ends_with(format!("{S}target{S}debug", S = MAIN_SEPARATOR).as_str())
+    || curr_dir.ends_with(format!("{S}target{S}release", S = MAIN_SEPARATOR).as_str())
+    || cfg!(target_os = "windows")
+  {
+    // running from the out dir or windows
+    return Ok(exe_dir.to_path_buf());
+  }
+
+  if cfg!(target_os = "linux") {
+    if curr_dir.ends_with("/data/usr/bin") {
+      // running from the deb bundle dir
+      Ok(exe_dir.join(format!("../lib/{}", app_name)))
+    } else {
+      // running bundle
+      Ok(PathBuf::from(format!("/usr/lib/{}", app_name)))
+    }
+  } else if cfg!(target_os = "macos") {
+    Ok(exe_dir.join("../Resources"))
+  } else {
+    Err(crate::Error::from("Unknown target_os"))
+  }
 }
