@@ -1,11 +1,11 @@
 mod cmd;
 
+#[cfg(not(any(feature = "dev-server", feature = "embedded-server")))]
+use std::path::PathBuf;
 use web_view::WebView;
 
-use crate::TauriResult;
-
 #[allow(unused_variables)]
-pub(crate) fn handle<T: 'static>(webview: &mut WebView<'_, T>, arg: &str) -> TauriResult<bool> {
+pub(crate) fn handle<T: 'static>(webview: &mut WebView<'_, T>, arg: &str) -> crate::Result<bool> {
   use cmd::Cmd::*;
   match serde_json::from_str(arg) {
     Err(_) => Ok(false),
@@ -113,7 +113,7 @@ pub(crate) fn handle<T: 'static>(webview: &mut WebView<'_, T>, arg: &str) -> Tau
   }
 }
 
-fn init() -> TauriResult<String> {
+fn init() -> crate::Result<String> {
   #[cfg(not(any(feature = "all-api", feature = "event")))]
   return Ok(String::from(""));
   #[cfg(any(feature = "all-api", feature = "event"))]
@@ -151,7 +151,7 @@ fn init() -> TauriResult<String> {
 }
 
 #[cfg(any(feature = "all-api", feature = "open"))]
-fn open_fn(uri: String) -> TauriResult<()> {
+fn open_fn(uri: String) -> crate::Result<()> {
   crate::spawn(move || {
     #[cfg(test)]
     assert!(uri.contains("http://"));
@@ -164,7 +164,7 @@ fn open_fn(uri: String) -> TauriResult<()> {
 }
 
 #[cfg(any(feature = "all-api", feature = "event"))]
-fn listen_fn(event: String, handler: String, once: bool) -> TauriResult<String> {
+fn listen_fn(event: String, handler: String, once: bool) -> crate::Result<String> {
   Ok(format!(
     "if (window['{listeners}'] === void 0) {{
       window['{listeners}'] = {{}}
@@ -198,19 +198,40 @@ fn load_asset<T: 'static>(
   asset_type: String,
   callback: String,
   error: String,
-) -> TauriResult<()> {
+) -> crate::Result<()> {
   let handle = webview.handle();
   crate::execute_promise(
     webview,
     move || {
-      let read_asset = crate::assets::ASSETS.get(&format!(
-        "{}{}{}",
-        env!("TAURI_DIST_DIR"),
-        if asset.starts_with("/") { "" } else { "/" },
-        asset
-      ));
-      if read_asset.is_err() {
-        return Err(r#""Asset not found""#.to_string());
+      let mut path = PathBuf::from(if asset.starts_with('/') {
+        asset.replacen("/", "", 1)
+      } else {
+        asset.clone()
+      });
+      let mut read_asset;
+      loop {
+        read_asset = crate::assets::ASSETS.get(&format!(
+          "{}/{}",
+          env!("TAURI_DIST_DIR"),
+          path.to_string_lossy()
+        ));
+        if read_asset.is_err() {
+          match path.iter().next() {
+            Some(component) => {
+              let first_component = component.to_str().expect("failed to read path component");
+              path = PathBuf::from(path.to_string_lossy().replacen(
+                format!("{}/", first_component).as_str(),
+                "",
+                1,
+              ));
+            }
+            None => {
+              return Err(format!("Asset '{}' not found", asset).into());
+            }
+          }
+        } else {
+          break;
+        }
       }
 
       if asset_type == "image" {
@@ -222,19 +243,23 @@ fn load_asset<T: 'static>(
           "jpeg"
         };
         Ok(format!(
-          "`data:image/{};base64,{}`",
+          r#""data:image/{};base64,{}""#,
           ext,
           base64::encode(&read_asset.expect("Failed to read asset type").into_owned())
         ))
       } else {
         handle
           .dispatch(move |_webview| {
-            _webview.eval(
-              &std::str::from_utf8(&read_asset.expect("Failed to read asset type").into_owned())
-                .expect("failed to convert asset bytes to u8 slice"),
-            )
+            let asset_bytes = &read_asset.expect("Failed to read asset type").into_owned();
+            let asset_str =
+              &std::str::from_utf8(asset_bytes).expect("failed to convert asset bytes to u8 slice");
+            if asset_type == "stylesheet" {
+              _webview.inject_css(asset_str)
+            } else {
+              _webview.eval(asset_str)
+            }
           })
-          .map_err(|err| format!("`{}`", err))
+          .map_err(|err| crate::ErrorKind::Promise(format!(r#""{}""#, err)).into())
           .map(|_| r#""Asset loaded successfully""#.to_string())
       }
     },
