@@ -1,4 +1,6 @@
 use std::sync::Mutex;
+
+use lazy_static::lazy_static;
 use uuid::Uuid;
 use web_view::WebView;
 
@@ -20,7 +22,7 @@ pub fn generate() -> String {
       value: salt.to_string(),
       one_time: true,
     });
-  return salt.to_string();
+  salt.to_string()
 }
 
 pub fn generate_static() -> String {
@@ -32,7 +34,7 @@ pub fn generate_static() -> String {
       value: salt.to_string(),
       one_time: false,
     });
-  return salt.to_string();
+  salt.to_string()
 }
 
 pub fn is_valid(salt: String) -> bool {
